refactor: update remaining stale imports after error subpackage migration
Update validation_port, domain entity errors, and validation_rule
to use errors.base.* and the new concrete RuleViolated/ValidationFailed.
