+++ src/forging_blocks/application/ports/inbound/validation_port.py
@@ -12,7 +12,7 @@ Non-Responsibilities:
 from abc import abstractmethod
 from typing import Any
 
-from forging_blocks.foundation.errors.rule_violation_error import RuleViolationError
+from forging_blocks.foundation.errors.base.rule_violation_error import RuleViolationError
 from forging_blocks.foundation.ports import InboundPort
