+++ src/forging_blocks/domain/errors/draft_entity_is_not_hashable_error.py
@@ -2,11 +2,11 @@
 
 from typing import Self
 
+from forging_blocks.foundation.errors.base.error import Error
 from forging_blocks.foundation.errors.core import ErrorMessage
-from forging_blocks.foundation.errors.error import Error
 
 
-class DraftEntityIsNotHashableError(Error[dict[str, object]]):
+class DraftEntityIsNotHashableError(Error[object]):
     """Raised because draft entities are not hashable."""
 
     @classmethod
