+++ src/forging_blocks/domain/errors/entity_id_deletion_error.py
@@ -1,10 +1,10 @@
 """Module defining the EntityIdDeletionError exception."""
 
+from forging_blocks.foundation.errors.base.error import Error
 from forging_blocks.foundation.errors.core import ErrorMessage, ErrorMetadata
-from forging_blocks.foundation.errors.error import Error
 
 
-class EntityIdDeletionError(Error[dict[str, object]]):
+class EntityIdDeletionError(Error[str]):
     """Raised when there is an attempt to delete an entity's identifier."""
 
     def __init__(self, class_name: str) -> None:
@@ -17,7 +17,7 @@ class EntityIdDeletionError(Error[dict[str, object]]):
         message = ErrorMessage(
             f"Cannot delete 'id' of {class_name} as it defines the entity's identity."
         )
-        metadata = ErrorMetadata[dict[str, object]](
+        metadata = ErrorMetadata(
             {
                 "class_name": class_name,
                 "attribute_name": "id",
