+++ src/forging_blocks/foundation/rules/validation_rule.py
@@ -3,7 +3,7 @@
 from abc import ABC, abstractmethod
 from typing import Any
 
-from forging_blocks.foundation.errors.rule_violation_error import RuleViolationError
+from forging_blocks.foundation.errors.base.rule_violation_error import RuleViolationError
 
 
 class ValidationRule(ABC):
